test: factor out exports module
