+++ packages/swing-store/test/test-repair-metadata.js
@@ -8,7 +8,7 @@ import sqlite3 from 'better-sqlite3';
 
 import { importSwingStore } from '../src/index.js';
 
-import { makeExporter, buildData } from './test-import.js';
+import { makeExporter, buildData } from './exports.js';
 import { tmpDir } from './util.js';
 
 test('repair metadata', async t => {
